Add a guard to avoid infinite loops.
This is an “interesting” one because the original
Implementation on root of lisp doesn’t have it.
But this definitely causes a infinite loop on some
of the white paper’s examples.

So I will be assuming that I probable messed up
something, but still I prefer to have it working
with this very mild discrepancy than to have it
not working at all.